Updates new classes

+++ flink-runtime/src/main/java/org/apache/flink/runtime/leaderelection/DefaultLeaderElection.java
@@ -58,7 +58,9 @@ class DefaultLeaderElection implements LeaderElection {
 
     @Override
     public void close() throws Exception {
-        parentService.remove(leaderContender);
-        leaderContender = null;
+        if (leaderContender != null) {
+            parentService.remove(leaderContender);
+            leaderContender = null;
+        }
     }
 }
